Make sendMessage add notes instead
Delete DisplayMessageActivity; change notePosition floats to ints;
Make sendMessage add notes instead; and sort notes by their height.

+++ app/src/main/java/ca/eyqs/kanon/NotePosition.java
@@ -1,16 +1,16 @@
 package ca.eyqs.kanon;
 
 public class NotePosition {
-    public float x;
-    public float y;
-    public NotePosition(float position, float height) {
+    public int x;
+    public int y;
+    public NotePosition(int position, int height) {
         x = position;
         y = height;
     }
-    public float getX() {
+    public int getX() {
         return x;
     }
-    public float getY() {
+    public int getY() {
         return y;
     }
 }
